fix: rename items and fix pr review comments
