+++ packages/desktop/views/dashboard/drawers/network-config/NetworkConfigRouterView.svelte
@@ -4,7 +4,7 @@
     import { NetworkConfigRoute, networkConfigRoute, NetworkConfigRouter, networkConfigRouter } from '../'
     import {
         AddCustomNetworkDrawer,
-        AddNetworkDrawer,
+        ManageNetworksDrawer,
         NetworkDepositAddressDrawer,
         NetworkInformationDrawer,
         ConfirmLedgerEvmAddressDrawer,
@@ -47,8 +47,8 @@
     <NetworkDepositAddressDrawer {drawerRouter} />
 {:else if $networkConfigRoute === NetworkConfigRoute.ConfirmLedgerEvmAddress}
     <ConfirmLedgerEvmAddressDrawer {drawerRouter} />
-{:else if $networkConfigRoute === NetworkConfigRoute.AddChain}
-    <AddNetworkDrawer {drawerRouter} />
+{:else if $networkConfigRoute === NetworkConfigRoute.ManageNetworks}
+    <ManageNetworksDrawer {drawerRouter} />
 {:else if $networkConfigRoute === NetworkConfigRoute.AddCustomChain}
     <AddCustomNetworkDrawer {drawerRouter} />
 {/if}
