+++ packages/desktop/views/dashboard/drawers/network-config/network-config-route.enum.ts
@@ -1,5 +1,5 @@
 export enum NetworkConfigRoute {
-    AddChain = 'addChain',
+    ManageNetworks = 'manageNetworks',
     AddCustomChain = 'addCustomChain',
     ChainDepositAddress = 'chainDepositAddress',
     ChainInformation = 'chainInformation',
