+++ packages/desktop/views/dashboard/drawers/network-config/views/ManageNetworksDrawer.svelte
@@ -12,7 +12,7 @@
     import { activeProfile } from '@core/profile/stores'
     import { Router } from '@core/router'
     import { NetworkAvatar } from '@ui'
-    import { NetworkConfigRoute } from '../'
+    import { NetworkConfigRoute } from '..'
 
     export let drawerRouter: Router<NetworkConfigRoute>
 
@@ -34,13 +34,13 @@
     }
 
     async function onContinueClick(): Promise<void> {
-        const promises = Object.keys(selectedChains).map(async (chainId) => {
-            if (!selectedChains[chainId] || $activeProfile?.evmNetworks.some((network) => network.id === chainId)) {
+        const promises = Object.keys(selectedChains).map(async (networkId) => {
+            if (!selectedChains[networkId] || $activeProfile?.evmNetworks.some((network) => network.id === networkId)) {
                 return
             }
 
-            const evmNetworkConfiguration = KNOWN_EVM_NETWORKS_CONFIGURATIONS.find(
-                (networkConfiguration) => networkConfiguration.chainId === chainId
+            const evmNetworkConfiguration = networkConfigurations.find(
+                (networkConfiguration) => networkConfiguration.id === networkId
             )
 
             if (!evmNetworkConfiguration) {
@@ -56,7 +56,7 @@
     }
 </script>
 
-<DrawerTemplate title={localize('views.dashboard.drawers.networkConfig.addChain.title')} {drawerRouter}>
+<DrawerTemplate title={localize('views.dashboard.drawers.networkConfig.manageNetworks.title')} {drawerRouter}>
     <known-chains class="h-full flex flex-col gap-4 px-6">
         {#each networkConfigurations as chainConfiguration}
             <Tile border>
@@ -72,8 +72,13 @@
                 </div>
             </Tile>
         {/each}
-        {#if Platform.isFeatureFlagEnabled('network.config.addChain.customChain')}
-            <Button variant="text" text="Add custom chain" icon={IconName.Plus} on:click={onAddCustomChainClick} />
+        {#if Platform.isFeatureFlagEnabled('network.config.manageNetworks.customChain')}
+            <Button
+                variant="text"
+                text={'views.dashboard.drawers.networkConfig.addChain.title'}
+                icon={IconName.Plus}
+                on:click={onAddCustomChainClick}
+            />
         {/if}
     </known-chains>
     <div slot="footer" class="flex justify-center">
