+++ packages/desktop/views/dashboard/drawers/network-config/views/index.ts
@@ -1,5 +1,5 @@
 export { default as AddCustomNetworkDrawer } from './AddCustomNetworkDrawer.svelte'
-export { default as AddNetworkDrawer } from './AddNetworkDrawer.svelte'
+export { default as ManageNetworksDrawer } from './ManageNetworksDrawer.svelte'
 export { default as NetworkDepositAddressDrawer } from './NetworkDepositAddressDrawer.svelte'
 export { default as NetworkInformationDrawer } from './NetworkInformationDrawer.svelte'
 export { default as ConfirmLedgerEvmAddressDrawer } from './ConfirmLedgerEvmAddressDrawer.svelte'
