+++ packages/shared/src/lib/features/interfaces/network-features.interface.ts
@@ -2,7 +2,7 @@ import { IFeatureFlag } from './feature-flag.interface'
 
 export interface INetworkFeatures extends IFeatureFlag {
     config: IFeatureFlag & {
-        addChain: IFeatureFlag & {
+        manageNetworks: IFeatureFlag & {
             customChain: IFeatureFlag
         }
         removeNetwork: IFeatureFlag
