+++ packages/shared/src/locales/en.json
@@ -431,8 +431,11 @@
             },
             "drawers": {
                 "networkConfig": {
+                    "manageNetworks": {
+                        "title": "Manage networks"
+                    },
                     "addChain": {
-                        "title": "Add chain"
+                        "title": "Add custom chain"
                     },
                     "chain": {
                         "chainId": "Chain ID",
